# tiktok-bitrix-sync/src/modules/tiktok/controllers/tiktok-webhook.controller.ts
import { Controller, Post, Body, UseGuards, Req } from '@nestjs/common';
import { ApiTags, ApiOperation, ApiResponse, ApiHeader } from '@nestjs/swagger';
import { TikTokWebhookService } from '../services/tiktok-webhook.service';
import { TikTokWebhookDto } from '../dto/tiktok-webhook.dto';
import { TikTokSignatureGuard } from '../../../common/guards/tiktok-signature.guard';
import { Public } from '../../../common/decorators/api-key.decorator';
import { API_CONSTANTS } from '../../../common/constants/api.constants';

@ApiTags('Webhooks')
@Controller('webhooks/tiktok')
export class TikTokWebhookController {
  constructor(private readonly webhookService: TikTokWebhookService) {}

  @Public()
  @UseGuards(TikTokSignatureGuard)
  @Post('leads')
  @ApiOperation({ summary: 'Receive TikTok Lead Generation webhook events' })
  @ApiHeader({
    name: API_CONSTANTS.TIKTOK_SIGNATURE_HEADER,
    description: 'HMAC-SHA256 signature generated by TikTok using your secret token',
    required: false,
  })
  @ApiResponse({ status: 200, description: 'Webhook event received and accepted' })
  @ApiResponse({ status: 401, description: 'Invalid or missing signature' })
  async handleLeadWebhook(@Body() dto: TikTokWebhookDto, @Req() req: any) {
    return this.webhookService.ingestWebhook(dto, req.body);
  }
}